DON'T MERGE seekPOSTorGET returns only ascii chars
This is just a suggestion on how to deal with non-ascii chars, to avoid the problems like the one we had on this night (vandalism?).

In this PR I make `seedPOST`, `seekGET`, and `seekPOSTorGET` return only valid ascii chars. When they find a non-ascii char it'll return a `.` (dot).

I'm highlighting the **DON'T MERGE** on the title because I'm not sure if all clients of these functions will work as expected with this behavior.