getting rid of "Free Genesis ROMs" section
IMHO this section doesn't fit very well on this page. People usually come here to download the emulators.

Also, I'm planning to add a "Free ROMs" page on the docs. This way the the community can update more dynamically, adding links to the ROMs and to the sets we already have on the site.