Added the ability to reorder site awards. There is a new button in the right container of the user settings page that will take the user to the new reorder site award page. This new page is set up just like the achievement reorder page when the user can enter the position number where they want the site award to display.
New site awards will be added to the bottom of the list and will have a display order number of the (max display order of the given user) + 1.

In order for the new site award reordering to work the 20190702_Add_DisplayOrder_to_SiteAwards.sql must be run to add the new DisplayOrder field to the SiteAwards table in the database. By default it will give all awards a DisplayOrder of 0 until the user goes in and changes them.

The code used to "squash" completed and mastered awards has been updated to work with the new display order and moved to a single location in user.php, there is no need to have the same code in three different locations.

Award tooltips now show the date they were awarded on.