Fix embedding of twitch videos/collections/clips
The replacement of twitch links with the embedded twitch player seemed to assume
that the text it operated on only contained the twitch link and nothing else. It
also assumed that if 'twitch.tv' was present in the text, the rest of the text
was a properly formed twitch link.

Fixed using a regex-based solution that can handle multiple twitch links and
does not mess with the rest of the text.

-----

A second problem was that twitch, youtube & imgur links were being embedded even
when they were introduced by the expansion of tags such as [user=x]. This caused
problems if a user had a twitch link in their motto, for example.

Fixed by embedding the links before expanding forum-specific tags, thus acting
only on the links present in the actual forum post.