Allow users to explicitly (un-)subscribe forum topics
Previously, users were always implicitly subscribed to new comments in forum
topics they had commented on. The only way out was to disable notifications
globally, thus unsubscribing from every forum topic.

Users can now explicitly unsubscribe from specific forum topics they are
currently subscribed to. This operation can be used to nullify the implicit
subscription that occurs when the user comments on a topic, which is still the
default behaviour.

Users are also free to re-subscribe to a previously unsubscribed topic, or even
subscribe to topics that they have not commented on.