do not notify user that originates comments from the Server user
When a developer creates/modifies an achievement, the operation is
logged in a comment posted by the special "Server" user. Such a comment
would usually originate a notification to the author of the achievement,
but there was special-case code in place to avoid that. However, PR 349
mistakenly disabled this safeguard due to a mistyped variable name.

Reintroduced the safeguard, but enhanced, because it, as was, had
another issue: only the achievement author was eligible for it. As a
result, any other developers that modified the achievement would always
receive the notification due to Server's log comment.

Now the safeguard is applied to the user that originated the comment
from Server, regardless of whether he's the achievement author or not.