fix email notifications for game ticket subscribers
PR #349 introduced the capability of developers subscribing to tickets
of a given game, including their creation. The ticket creation email
notifications (both for the author and subscribers) is sent from a
different piece of code than most other notifications. The code that
sent the email to subscribers had a variable typo that kept the emails
from going out. The notification for the achievement author was
unaffected.

While searching for this problem I also noticed that in some places the
user settings notification flag being used was incorrect, so I fixed
that as well.