Advanced Ticket Filtering
This feature add several different ways of filter tickets on the Ticket Manager page. Each filter can be individually toggled which allows for a good amount of custom filtering. Total open tickets and total filtered tickets will always display in the headers on the page.

The filters are broken into the following categories:

Ticket State - The state of the ticket
 * Open
 * Closed
 * Resolved

Report Type - Why is this ticket being written
 * Triggered at wrong time
 * Doesn't Trigger

MD5 - Does the ticket include a MD5
 * Contains MD5 - (e.g. 'MD5: e9b45d6455e0753b8e0e825a36458253')
 * MD5 Unknown - (e.g. 'MD5: Unknown' or 'MD5:' is not in the notes at all)

Emulator - Does the ticket specify an emulator
 * RA Emulator - (e.g. 'Emulator: RANES')
 * RetroArch - Defined - (e.g. 'Emulator: RetroArch (Mupen64Plus-Next OpenGL)')
 * RetroArch - Undefined - (e.g. 'Emulator: RetroArch ()')
 * Emulator Unknown - Undefined - (e.g. 'Emulator: Something' or 'Emulator:' is not in the notes at all)