Output sanitization
Now, if you would please just let us know via Discord or a direct
message to RAdmin on-site, or post a pull request fixing whatever you
found instead of posting to bountysource for points to increase your
hax0r-e-eppen that'd be nice. We were well aware that there are XSS and
SQL injection vectors. Thank you.