chore: php8.2 upgrade
We're already running PHP 8.2 and Ubuntu 22.04 on the server.
This update reflect those changes in the application.
To be able to upgrade to Laravel 10 and use Enums we need 8.1+.
8.0 will be EOL (end of life) in two months.

Note: composer update to pull in PHP 8.2 dependencies will follow in a
separate PR because it will require some changes.

Please update your local environment:
- Ubuntu 20.04 in VM: dist upgrade
- XAMPP/LAMPP/WAMPP: re-install/update
- Docker: `sail up`
- Valet: `valet use 8.2` or `valet isolate php@8.2`