fix(game): remediate issue with '-' sort orders on some game pages (#4306)
http://localhost:64000/game/29902

On some game pages, both ascending and descending sorts produce
identical results. On the linked page, "Display order (first)" and
"Display order (last)" do the same thing in `master`. This is happening
because we're accidentally not sending down a field we need to the UI.

Side note: the fact no one has reported this as a bug leads me to
believe we should cut some of these sort orders for the sake of
simplicity. That may be worth doing in a follow-up.