fix(user): remediate issue with 'Hide completed games' toggle (#4312)
Resolves #4309.

http://localhost:64000/user/Kinglink

On `master`, enabling the "Hide completed games" toggle causes 0 rows to
appear, because the current set of 200 rows returned represents a
complete list of fully completed games.

Now, we fetch two lists on page load: the current list of 200 games, and
a list of up to 200 incomplete games. Based on the toggle status, we
toggle between which list of rows is being displayed.

The memory bump on the page is modest - I'm at ~45MB per hit locally. I
was originally set on not fixing this because I don't want to introduce
a new HTTP call to resolve the issue that will need to be thrown away
when `user2` rolls around, but this solution feels like a nice happy
medium.

<img width="378" height="473" alt="Screenshot 2025-12-26 at 8 38 47 AM"
src="https://github.com/user-attachments/assets/3ff5fd9f-54fb-492e-b14d-318b42c0d91a"
/>