chore: modernize the Friends table schema (#4314)
This PR modernizes the Friends table by renaming it to `user_relations`
and updating all legacy column names:
```
Friendship -> status
Created -> created_at
Updated -> updated_at
```
Additionally, the `UserRelationship` enum has been converted from an
abstract class with integer constants to a native PHP enum with string
backing:
* `UserRelationStatus`: `-1/0/1` -> `blocked/not_following/following`

The enum includes a `toLegacyInteger()` method for V1 API backwards
compatibility.

Similar to a few other of these maintenance PRs, **the migration mutates
table data**. This is typically undesirable, however I feel it's
warranted here to mitigate the need for multiple migrations and/or
SQL-on-deploy. This is a one-and-done.